autodroid-container/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
appium-python-client==2.11.1
asyncio==3.4.3
alembic==1.13.1
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
psutil==5.9.8
pytest==7.4.4
pytest-asyncio==0.23.3
zenoh==0.10.0
